Add initial draft of Jill's calendar email for June 2025
This commit introduces a new markdown file containing an email from Theo to Jill regarding the Heritage Happenings calendar. The email discusses the diminishing quality of the calendar and activities program, as well as the lack of interest from recipients. It also reflects on the responsibilities of sending out the calendar and invites Jill's thoughts on the matter.